+++ Cargo.toml
@@ -8,6 +8,7 @@ daft-dsl = {path = "src/daft-dsl", default-features = false}
 daft-execution = {path = "src/daft-execution", default-features = false}
 daft-io = {path = "src/daft-io", default-features = false}
 daft-json = {path = "src/daft-json", default-features = false}
+daft-local-execution = {path = "src/daft-local-execution", default-features = false}
 daft-micropartition = {path = "src/daft-micropartition", default-features = false}
 daft-parquet = {path = "src/daft-parquet", default-features = false}
 daft-plan = {path = "src/daft-plan", default-features = false}
@@ -31,6 +32,7 @@ python = [
   "daft-csv/python",
   "daft-dsl/python",
   "daft-execution/python",
+  "daft-local-execution/python",
   "daft-io/python",
   "daft-json/python",
   "daft-micropartition/python",
@@ -93,6 +95,7 @@ members = [
   "src/common/system-info",
   "src/daft-core",
   "src/daft-execution",
+  "src/daft-local-execution",
   "src/daft-io",
   "src/daft-parquet",
   "src/daft-csv",
