--- src/daft-execution/src/simple/mod.rs
@@ -1,5 +1,0 @@
-pub mod common;
-pub mod intermediate_op;
-pub mod pipeline;
-pub mod sinks;
-pub mod visitor;
