+++ src/daft-execution/src/stage/run.rs
@@ -19,27 +19,8 @@ use crate::{
         Executor,
     },
     partition::PartitionRef,
-    simple::{pipeline::execute_pipelines, visitor::PhysicalToPipelineVisitor},
     stage::Stage,
 };
-
-pub fn run_local_simple(
-    query_stage: &QueryStageOutput,
-    psets: HashMap<String, Vec<Arc<MicroPartition>>>,
-) -> DaftResult<Box<dyn Iterator<Item = DaftResult<Arc<MicroPartition>>> + Send>> {
-    let (physical_plan, _is_final) = match query_stage {
-        QueryStageOutput::Partial { physical_plan, .. } => (physical_plan.as_ref(), false),
-        QueryStageOutput::Final { physical_plan, .. } => (physical_plan.as_ref(), true),
-    };
-    let mut pipeline_visitor = PhysicalToPipelineVisitor {
-        pipelines: Vec::new(),
-        psets: psets.clone(),
-    };
-    Arc::new(physical_plan.clone()).visit(&mut pipeline_visitor)?;
-    let pipelines = pipeline_visitor.pipelines;
-    let result = execute_pipelines(pipelines)?;
-    Ok(Box::new(result.into_iter().map(Ok)))
-}
 
 /// Run a stage locally and synchronously, with all tasks executed serially.
 pub fn run_local_sync(
